requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
azure-core
azure-ai-documentintelligence
flask
python-dotenv
PyPDF2
ultralytics
pdf2image
pillow
pymupdf

python-dotenv
opencv-python-headless
# opencv-python
numpy
